Refactor test utils to prevent flakiness
