+++ posthog/temporal/tests/utils/events.py
@@ -4,15 +4,29 @@ import json
 import uuid
 import random
 import typing
-import asyncio
 import datetime as dt
 import itertools
 
 import aiohttp.client_exceptions
+from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_random_exponential
 
 from posthog.models.raw_sessions.sessions_v2 import RAW_SESSION_TABLE_BACKFILL_SELECT_SQL
-from posthog.temporal.common.clickhouse import ClickHouseClient
+from posthog.temporal.common.clickhouse import ClickHouseClient, ClickHouseError
 from posthog.temporal.tests.utils.datetimes import date_range
+
+
+@retry(
+    retry=retry_if_exception_type(
+        (aiohttp.client_exceptions.ClientOSError, aiohttp.client_exceptions.ServerDisconnectedError, ClickHouseError)
+    ),
+    # on attempts expired, raise the exception encountered in our code, not tenacity's retry error
+    reraise=True,
+    wait=wait_random_exponential(multiplier=0.2, max=3),
+    stop=stop_after_attempt(5),
+)
+async def execute_query(clickhouse_client: ClickHouseClient, query: str, *data):
+    """Try to prevent flakiness in CI by retrying the query if it fails."""
+    return await clickhouse_client.execute_query(query, *data)
 
 
 class EventValues(typing.TypedDict):
@@ -107,60 +121,45 @@ async def insert_event_values_in_clickhouse(
     client: ClickHouseClient, events: list[EventValues], table: str = "sharded_events", insert_sessions: bool = False
 ):
     """Execute an insert query to insert provided EventValues into sharded_events."""
-    max_attempts = 3
-    attempt = 1
-    backoff = 2
-
-    while True:
-        try:
-            await client.execute_query(
-                f"""
-            INSERT INTO `{table}` (
-                uuid,
-                event,
-                timestamp,
-                _timestamp,
-                person_id,
-                team_id,
-                properties,
-                elements_chain,
-                distinct_id,
-                inserted_at,
-                created_at,
-                person_properties
+    await execute_query(
+        client,
+        f"""
+    INSERT INTO `{table}` (
+        uuid,
+        event,
+        timestamp,
+        _timestamp,
+        person_id,
+        team_id,
+        properties,
+        elements_chain,
+        distinct_id,
+        inserted_at,
+        created_at,
+        person_properties
+    )
+    VALUES
+    """,
+        *[
+            (
+                event["uuid"],
+                event["event"],
+                event["timestamp"],
+                event["_timestamp"],
+                event["person_id"],
+                event["team_id"],
+                json.dumps(event["properties"]) if isinstance(event["properties"], dict) else event["properties"],
+                event["elements_chain"],
+                event["distinct_id"],
+                event["inserted_at"],
+                event["created_at"],
+                json.dumps(event["person_properties"])
+                if isinstance(event["person_properties"], dict)
+                else event["person_properties"],
             )
-            VALUES
-            """,
-                *[
-                    (
-                        event["uuid"],
-                        event["event"],
-                        event["timestamp"],
-                        event["_timestamp"],
-                        event["person_id"],
-                        event["team_id"],
-                        json.dumps(event["properties"])
-                        if isinstance(event["properties"], dict)
-                        else event["properties"],
-                        event["elements_chain"],
-                        event["distinct_id"],
-                        event["inserted_at"],
-                        event["created_at"],
-                        json.dumps(event["person_properties"])
-                        if isinstance(event["person_properties"], dict)
-                        else event["person_properties"],
-                    )
-                    for event in events
-                ],
-            )
-            break  # Success, exit the loop
-        except (aiohttp.client_exceptions.ClientOSError, aiohttp.client_exceptions.ServerDisconnectedError):
-            if attempt >= max_attempts:
-                raise
-
-            attempt += 1
-            await asyncio.sleep(backoff)
-            backoff = backoff**2
+            for event in events
+        ],
+    )
 
 
 async def insert_sessions_in_clickhouse(client: ClickHouseClient, table: str = "sharded_events"):
@@ -171,10 +170,13 @@ async def insert_sessions_in_clickhouse(client: ClickHouseClient, table: str = "
             "`$session_id`", "JSONExtractString(properties, '$session_id')"
         )
 
-    await client.execute_query(f"""
+    await execute_query(
+        client,
+        f"""
     INSERT INTO raw_sessions
     {generate_sessions_query}
-    """)
+    """,
+    )
 
 
 async def generate_test_events_in_clickhouse(
