+++ products/batch_exports/backend/tests/temporal/conftest.py
@@ -22,12 +22,12 @@ from posthog.temporal.common.clickhouse import ClickHouseClient
 from posthog.temporal.common.client import connect
 from posthog.temporal.common.logger import configure_logger
 from posthog.temporal.tests.utils.events import generate_test_events_in_clickhouse
-from posthog.temporal.tests.utils.persons import (
+
+from products.batch_exports.backend.temporal.metrics import BatchExportsMetricsInterceptor
+from products.batch_exports.backend.tests.temporal.utils.persons import (
     generate_test_person_distinct_id2_in_clickhouse,
     generate_test_persons_in_clickhouse,
 )
-
-from products.batch_exports.backend.temporal.metrics import BatchExportsMetricsInterceptor
 
 
 @pytest.fixture(scope="package", autouse=True)
