+++ products/batch_exports/backend/tests/temporal/destinations/bigquery/utils.py
@@ -13,7 +13,7 @@ from posthog.temporal.common.clickhouse import ClickHouseClient
 from products.batch_exports.backend.temporal.destinations.bigquery_batch_export import bigquery_default_fields
 from products.batch_exports.backend.temporal.record_batch_model import SessionsRecordBatchModel
 from products.batch_exports.backend.temporal.spmc import Producer, RecordBatchQueue
-from products.batch_exports.backend.tests.temporal.utils import get_record_batch_from_queue
+from products.batch_exports.backend.tests.temporal.utils.records import get_record_batch_from_queue
 
 SKIP_IF_MISSING_GOOGLE_APPLICATION_CREDENTIALS = pytest.mark.skipif(
     "GOOGLE_APPLICATION_CREDENTIALS" not in os.environ,
