+++ products/batch_exports/backend/tests/temporal/destinations/redshift/test_copy_activity.py
@@ -8,10 +8,6 @@ from psycopg import sql
 
 from posthog.batch_exports.service import BatchExportInsertInputs, BatchExportModel, BatchExportSchema
 from posthog.temporal.tests.utils.events import generate_test_events_in_clickhouse
-from posthog.temporal.tests.utils.persons import (
-    generate_test_person_distinct_id2_in_clickhouse,
-    generate_test_persons_in_clickhouse,
-)
 
 from products.batch_exports.backend.temporal.destinations.redshift_batch_export import (
     AWSCredentials,
@@ -33,6 +29,10 @@ from products.batch_exports.backend.tests.temporal.destinations.redshift.utils i
     assert_clickhouse_records_in_redshift,
     delete_all_from_s3_prefix,
     has_valid_credentials,
+)
+from products.batch_exports.backend.tests.temporal.utils.persons import (
+    generate_test_person_distinct_id2_in_clickhouse,
+    generate_test_persons_in_clickhouse,
 )
 
 pytestmark = [
