+++ products/batch_exports/backend/tests/temporal/destinations/snowflake/test_activity_e2e.py
@@ -14,10 +14,6 @@ from django.test import override_settings
 
 from posthog.batch_exports.service import BatchExportModel, BatchExportSchema
 from posthog.temporal.tests.utils.events import generate_test_events_in_clickhouse
-from posthog.temporal.tests.utils.persons import (
-    generate_test_person_distinct_id2_in_clickhouse,
-    generate_test_persons_in_clickhouse,
-)
 
 from products.batch_exports.backend.temporal.destinations.snowflake_batch_export import (
     SnowflakeInsertInputs,
@@ -34,6 +30,10 @@ from products.batch_exports.backend.tests.temporal.destinations.snowflake.utils 
     SKIP_IF_MISSING_REQUIRED_ENV_VARS,
     TEST_MODELS,
     assert_clickhouse_records_in_snowflake,
+)
+from products.batch_exports.backend.tests.temporal.utils.persons import (
+    generate_test_person_distinct_id2_in_clickhouse,
+    generate_test_persons_in_clickhouse,
 )
 
 pytestmark = [
